macos: Try replacing the sleep with a call to get the mouse location
In https://github.com/enigo-rs/enigo/issues/182 the user noticed that a timing issue goes away when fetching the current mouse location. I am not sure if that is because it just takes enough time to execute that it fixes it and a sleep would be equally good.

The changes kinda work. Simulating input is magnitudes faster, but there are issues. It works fine for about 20 simulations, but if you simulate more, some are missing. If you simulate 500, about 380 are successfully entered. If you simulate 20 in a loop it works fine. If you simulate 1000 about 800 are successful. Its very weird, because not only are some events missing, it also enters weird stuff. It entered "again" for example.