tests: Don't wait for a message from the browser before the next input is simulated
A lot of the bugs are caused by sending the input in short sequence. It is impossible to detect them if we always wait for the input to take effect. We now simulate the input and keep track of the events that should have happened. Only after everything was simulated and the Enigo struct was dropped do we check if all expected events were created by the browser. It is important to drop the Enigo struct before checking the events because some bugs were caused by dropping the struct before all events were handles by the OS.